add support for tile filtering - thanks to Rotoflex for the idea !
Currently, only simple inversion is done with the help of PIL. The tiles look like negative film as a result. For best night-time usage, ose with layers that are mostly white as all white areas will be inverted to solid black. Known issue: doesn't seem to work very well with transparent layers.