change watch() argument order; add runNow
The modrana.watch() method now takes callback arguments as a list and also accepts the runNow optional argument. If runNow is set, the callback is run once when it is registered. This is a frequent use-case for many modules, that configure their state based on keys on options on startup. With runNpw, they don't have to have separate startup & value changed code anymore.