load the location page and the speed page dynamically
Like this, the pages are not loaded until needed and as they react on position changes they react only when loaded/visible - this simplifies the page design as the pages can use regular property bindings and don't have to check if they are visible or not. The bindings just won't fire if the page sis not visible as the page would not be loaded.