Add a modrana context property, fix theme errors at startup
The modrana context property exposes core modRana functionality to QML. Currently, it has mode and theme properties that can be used for bindings and also to set the mode and theme in the modRana core. also all rWin.theme calls were replaced by modrana.theme, which fixes quite a few errors reported in the startup log.