Add support for batch tile update
The redownload toggle button in the batch "Edit" submenu now has a new "update"
state. When in this state, only tiles that are already available will be downloaded.
In short, it will update all already downloaded tiles for the current layer in the given area.

Also using a big area, this can be used a rudimentary tile db update tool.