Add support for conditional callbacks
The modRanaThread class now has a target variable and callback property. Target holds the method that will be called and  callback is an optional callback handler. If callback is set to None or False when the thread is running, the callback would not be called. This enables running various tasks (such as online API requests) in threads and being able to cancel them (so that they don't return any unexpected results) while the thread is running.