Add direction to point arrow
In menu->Info->Direction to point, there is a new screen with an arrow, changing the direction to a given point in respect to the current direction of travel.
If the target point is directly in front of the direction of travel, the arrow should point up, if behind, it should point down and if it is somewhere else, it should point accordingly.
ModRana remembers the las used coordinates, which can be cleared using the "clear" button.