Get connectivity state from conic on the N900
Looks like there was a typo in the callback preventing it from working. :)
Also, if the connection is requested directly from conic,
the callback is immediately called, so connectivity state can be updated.
The callback is also called at-once once registered, this means
that connectivity can be tracked through the whole application run
and using the portable implementation from the baseclass is not needed.
As the portable implementation might not cover all Internet connectivity options
on the N900, this should provide more robust connectivity tracking.
We can also get rid of the autoconnect shell script when calling conic directly
to trigger autoconnect.