Fix ComboBox to actually work on Silica
I've actually writen it some time before I started running modRana
on Silica, so it is rather good that just very few changes were needed.
Also #unlike the Controls combobox that automatically assigns its size,
the silica combobox is expected to fill the width of the screen,
so we need to anchor to our parents left & right anchors.