Add support for tile scaling
Tile scaling makes it possible to scale tiles on a given
zoomlevel and make them bigger or smaller.
Currently only up-scaling is possible, which makes
possible to enlarge features on the tiles at the cost
of blurring and resize artifacts.
Down-scaling while doable is currently
not possible (int scale factor) as it would
mean more tiles on the screen & doesn't
really bring much that can't be achieved
by just using a higher zoomlevel.