Properly handle hexadecimal color specification in GTK GUI
GTK2 for some reason can't parse alpha value from hexadecimal color
specification, so try to split ty manually if parsing blows up.

Also fix a missing # in hexadecimal color specification in the
Silica theme and add a note bout how important it is to include
# in hexa color specs & that numeric alpha value will be ignored
if alpha is part of the spec string.