Add a unfinished page debugging property
The root window showUnfinishedPages makes it possible to show unfinished
and debug them, but as it is false by default, unfinished pages will not be
shown by default (actually not the page itself, but rather the
button leading to it is hidden).

And of course once a given page is done, this condition will be removed
and the page will be accessible in the same way as other pages.