Fix Map1.eu not showing up
It was not showing up as it didn't have any group assigned.
While the original idea was to be show layers that don't
have any group assigned together with map layer groups,
the current map layer selectors in GTK, Qt 4 and Qt 5 GUI just don't
do that. So let's just move Map1.eu to a group corresponding to
Europe (as this is what it covers) so it is shown properly. :)