Add a custom thread pool implementations
Currently we have two thread pool implementations differentiated
by the work queue structure they use - one of them uses a priority queue,
which makes it possible to shut it down either immediately
(even if there are work item pending) or once all work is complete.
The  other thread pool uses a stack so that it can handle recently
added. The stack can also "leak" the bottom most item if it becomes full.
The item is returned to the caller of the submit method.

Both stacks also support discarding requests that were in the queue for
longer than a given timeout.