Add toggle button for setting local search radius
Like this the local search radius can easily switched without going to Options.

NOTE: The toggle button is only synced with radius from Option on start
and might show old values if the radius is changed in options. But it still
works when it is clicked and sets the value it is shoewing.