Include old Python 2.5 compatible Urllib 3
The new Urllib 3 is too know to easily convert to be both work
with Python 2.6+/3 and Python 2.5, so let's just include the old
working version and use it with Python <= 2.5 and make newer Python
versions will use the new one.

Also move the bundled Urllib 3 module to the core/bundle directory
as the first step in the initiative to have all bundled stuff in a single
directory so it can be easily removed on platforms that don't like bundling
or where it is not needed, such as on Fedora.