Use asynchronous tile downloads
Previously, tiles were downloaded by the asynchronous tile loading thread.
Unfortunately it looks like there is only one such thread and so even if one tile
download stalls, other downloads that might be quick are also blocked.

This was changed so that if a tile is not found by the image provider a special 1x1
pixel tile is returned and a download request is queued. Meanwhile the tile gets
"notified" by the 1x1 image and disables caching (so that the 1x1 image is not
cached), sets source to "" and connects to the "tile downloaded" signal.
Once the tile is downloaded, it triggers a Python signal which in turn sends a
message to QML. This message triggers a QML signal.
All tiles waiting for download will get notified, but only one of them will match
its tile id - the tile that has been downloaded. It will then enable disconnect
from the signal, enable caching and set it's source as it knows it's tile file
is now available.

Simple, eh ? :)

(If only the QtQuick Image loading API had some page-fault like hooks or a callback
based interface - call this with this id once the image is available - or something
similar.)