Final QML GUI asynchronous tile download tweaks
Well, that was tedious! :D It all would be so much easier if the QML image loading API was really asynchronous
and/or if it was possible to directly manipulate the image cache. Anyway, it seem to work now. :P
Hopefully it also continues working in the future. :D

What was done:
* tile download requests are sent after the 1x1 "page fault" tile is received, this seem to fix a race condition where
  the 1x1 tile was received after the tile was already downloaded
* don't set the source to "" after page fault - seems to introduce yet another race condition that might also overwrite
  the loaded image - just set opacity to 0.0 instead
* lots and lots of fiddling with the "cache" property
* support for handling tile download errors, dropped download requests and retrying
* increased the request queue as it turns out it was set to be too small
* improved tile handling debugging info