Use the getPage function from rWin
The function caches the pages it provides so
that the search pages should be parsed only once,
but also hold a reference to the page object.
This means they will not be garbage collected
after they are popped from the page stack together
with their result models, so the result model can be
used to show the results on the map without vanishing.