Show the search results on the map once a result is clicked
Once a result is clicked firs highlight the result item in the
data model and then assign the data model to the marker display
element in the main map pinch map. Like this all results will
be displayed and the results that was clicked will be highlighted
by a red border ("normal" result points have a blue border.