Add an optional callback to the options set() function
As there might be watches on some keys in the persistent options directory,
we might want to run a callback once the watches have been run, which is after
the normal set function has returned. So the set() function now has a new optional
argument called "callback" that can contain a callback function to be called once
the set() function on the Python side has done its work.