Make content assignment Qt 5.2+ (?) compatible
For some reason, on Jolla it is possible to do this in BasePAge instances:

content {
    Item {}
}

But it is an error when running with Qt 5.3 and I'm almost certain that
also with 5.2. As we want the modRana GUI to be robust and multi-platform,
we need to use the method that works on both Jolla and normal Qt, even if it
might be a bit less convenient in some places.