Remove the old config module
The original Rana idea was that like in some window managers or in Emacs,
configuration files would just be valid Python code. This unfortunately
turned to be too fragile, so modRana got support fo classical ini-style/key-value
(but with nesting support) configuration files. Also more and more of the defaults
set at startup by the old config module moved to the constants core module.
With moving of the thread pool defaults to the constants core module the
old config module can be removed.