Make it possible to do automatic tile downloads with one thread only
This could be useful for resource constrained environments, for using tileservers
prone to banning parallel downloads and to potentially work around resource related
crashes like reported on the N900.

Also show a notification when changing the number of threads that to apply the setting
modRana needs to be restarted (it might be possible to remove this limitation in the future).