Make it possible to change the number of batch download threads
This can be used for example to just use a single thread when downloading from tilservers
that don't like parallel downloads or to use more threads to download from won tileservers
or other tileservers that allow massively parallel downloads.

Downloading in multiple threads can also be more demanding on the downloading device,
so users that might have issues wih modRana being slow or unresponsive (or crashing as
reported in some cases on the N900) might want to lower the number of threads used for
batch download.