Remove the old update() function an related machinery
Rana & modRana used to have a 100 ms (!) timer polling all (!) modules,
with the modules checking the clock to see if it is time to do something
or if a given poll call can be skipped. As you might imagine it was
horrendously inefficient (double so on a battery-powered device) and
thankfully has been replaced per-task timers long time ago.

Still, there were some remaining non-functional pieces (actually the
100 ms timer has still been enabled but errored out due to wrong
function signature) and there were now removed.