Add OptionsProperty to Qt 5 GUI
This new "backend" element gets a default value and name of a modRana persistent options key.
Once instantiated, it fetches the value of the key asynchronously, replacing the default value with
the value returned by the callback once done.

Another feature of OptionsProperty is that once its value is set, it is stored to the corresponding
options key on the Python side.

This makes it possible to provide the name of the key only once and eliminates needless boilerplate
code for getting and setting the key value.

Also, OptionsProperty has a built-in protection against the key being set until the key-getting callback finishes,
this should prevent race conditions where switches might rewrite the value of a key due to onChecked triggering
at switch instantiation time.