Use OptProp aliases instead of direct property access
Like this we don't have to remember which property is a "normal" property
and which is an object where we need to access it's value property.
Also apparently nested property lookup might have some cost while the
alias seems to be a direct reference and thus probably faster, which
might be important if the property is queried very often (tile debug texts for example).