Add screen blanking control support to Sailfish OS
The Qt 5 GUI on Sailfish OS can now keep the screen on as long
as needed! :) The screen blanking prevention is bound to modRana
being the currently active application, so it does not keep
the screen on if it is in background or another application is active.

And of course the possibility of the import failing is taken into account
and screen blanking control is automatically disable outside of Sailfish OS.