Fix a race condition preventing some tiles from working correctly
There was a race that could prevent them from working correctly by breaking their
initialization by clearing the layers data model when restoring saved layers.
By instead replacing the initial item we can prevent the race and everything
seems to work fine. :) Still, if it turns out that loadLayers() needs to be called
after initialization in the future, this would need to be extended to handle an
arbitrary number of existing layers when loading saved layers.