Add a new core module for qrc handling
This module handles various aspects of qrc usage, such as detecting
that modRana is running on top of qrc or unpacking of files that modRana
needs as normal files and not resources.

The module is currently used on the Android platform because modRana
is using here for effective deployment. This is mostly because on Android all installed
APKs are stored on the device, so any extracted files are effectively present twice &
require twice the storage. Using qrc also simplifies distribution and deployment,
as the Qt SDK handles qrc generation transparently and only a few files
need to be extracted on first start, the rest is transferred directly to
RAM on demand.