Add support for early debug logging on Android
Given android crazy architecture, much can go wrong early in modRanas
startup, so we need to be able to enable the log file as early as
possible. So if we are running on Android and the
/sdcard/modrana_debug_logs folder exists, enable the log file at once,
even if the log file is not enabled in options.

Also make sure that logging to file can't be started twice, to avoid
possible breakage if early debug logging is enabled together with
log file enabled from options. Furthermore early debug logs can't
be disabled once started - this prevents them from being disabled
from options by accident, but could theoretically prevent the log file
from being cleanly closed. This should however not be that important
for a debug log targeted primarily for startup monitoring.