Only add the qrc root to Python import path on Android
It seems to lead to weird crashes on PC and Sailfish if
combined with "import io.thp.pyotherside 1.3" (with 1.0
import it works fine). But on Android we need to import 1.3
and it crashes with 1.0 import. :D
So just don't add "qrc:/" to Python import path outside
of Android for now. :)