Use wrapping functions for accessing icons
This makes it possible to get icons directly from qrc on Android,
saving space as we don't need to unpack the icons before use.

Technical note: All installed APKs are stored on Android, so anything that
is unpacked from the APK consumes a double amount of storage space as it is once
in the APK and once outside. So by not unpacking the icons and using them directly
from qrc which resides in the APK storage space is saved.