Initial PoC how routing might be attempted
This is an initial PoC attempt to do routing related stuff on the map,
currently it unconditionally places two areas "Buttons" on the map to
allow selection of routing start and destination points. It should paint
a line from start to destination on the map as an attempt to use the
canvas.

When selecting the destination point, it attempts to get a route. No
checks are done yet and nothing is done with the route.