Make it possible to enable back button even if not needed
Some users might want to have an onscreen back button even when the
platform provides an alternative mechanism for going to the previous page.

Sailfish OS is na example of such a platform. Android might be another if
it turns out to be possible to use the Android back button to go to the
previous page.