Fix POI detail menu breaking the GTK GUI once back is clicked
When a last visible POI has been restored on startup, then clicking on its label
and tne clicking on the back arrow would turn the screen black, breaking the GUI.
This happened because the POI menu structure has not been generated yet.

So fix this by marking the menu structure as "dirty" when the showPOI module is
initialized to trigger the POI menu structure regeneration before switching to the menu.