Add Qt 5 backend for the modRana cron module
The cron module in modRana serves a timer broker to other modRana
modules and makes it possible for the modRana Python core to shedule
repetetive tasks, such as saving track log increments to storage or
periodically checking if a route is still being followed.

The module currently has backends for using the GTK and Qt 4 main loops
and this commit adds a backend for Qt 5.

How does it work ?

Due to how PyOtherSide works, it is rather more complicated to get to
the Qt 5 main loop as no direct Qt 5 bindings are used and all
comuunication with Qt 5 goes through PyOtherside. To solve this issue,
a simple timeout broker has been added to the QML side and the Python
side cron backend communicates with it over pyotherside to add, remove
and modify individual timers. Then when the timers are running they
call a Python side entry point in the cron module with the id of the
timer that has triggered and the cron module then converts this call
to calling the callback that has been assigned to the given timer
when the timer has been registered.

Considering how crazy this sound (and kinda is! ;) ) it seems to work
just fine in practise. :)