Add initial track logging page for the Qt 5 GUI
General track logging works fine already, but the page is
for the moment marked as unfinished (and thus not visible by default)
until some of the few missing features (such as correctly loading
the last saved tracklog name) are implemented. Also some
nice icons for the menu with the Sailfish theme. :)