Extension and image type independent tile lookup
Return any suitable image for the given tile coordinates,
independently on file extension and image type.

For files the file extension specified by the tile layer is
tried first and if no such file exists or is not an image,
modRana iterates for all suitable files for the given coordinates
and returns the first image file it finds.

For sqlite tile storage modRana no longer includes the current extension
in the store database select, so tile data should be returned even if it's
extension does not match the current one. As x, y & z are used as a primary key,
we don't need to worry about checking multiple tiles for a given coordinate
combination. But modRana still checks the blob returned from sqlite and
logs a warning if it is not an image.

All in all this change should help both in case that non-image data got saved among
image files, but also when the default file extension changes for a layer
and previously stored tiles for the affected layer should no longer become inaccessible.
Any non image tile files found are now also logged, so it should be now easier to detect
and remove them.