Send track recording status updates by default
This work around a current bug in the UC Page when running
with the Controls backend - the isActive property is always true,
so the isActiveChanged signal is never triggered and the
status updates are never turned on or off.
But even once this is fixed, there should be no harm in enabling
sending of updates by default as they should be turned off once
the track recording page becomes inactive.