Move Qt 5 specific routing code to the Qt 5 GUI module
We can't import pyotherside from the route module as
that would break modRana on platforms where the Qt 5
GUI is not used (N900, some desktop platforms).
Also the amount of GUI specific code in the modRana
non-GUI modules should be minimized.

So move the Qt 5 specific routing code to the Qt 5 GUI module
and also skip some GTK related routing stuff when not running
with the GTK GUI.