Add a Popup component
This component is based on the Popup component from Mitakuuluu.
The Silica version is is basically the same as the Mitakuuluu
Popup, the Controls version has been modified to use the style
table and hardcoded default color instead of using a Silica
theme color. Mitakuuluu is licensed under WTFPL, so including
the Popup component in Universal Controls should be fine licensing
wise.