Signal based notifications
Replace the notification handling spaghetti code by
a single notificationTriggered signal in the modRana "kernel".
Notification connect to this signal and are notified when a notification
needs to be shown. The individual notification handlers can
decide if they connect to the signal or not themselves,
which scales much better than trying to centrally decide
what should be the current notification handler.