Comment out the verbose tile loading debugging label
Turns out the property changes have been triggering for the text property even
though the Label was invisible and as this was happening for every tile * number
of layers, it was actually creating a noticeable performance impact. So
noticeable that it could be clearly seen on the output of the QML profiler,
which is how I found about this.

So just comment it out for now and uncomment it when needed for debugging to eliminate the performance impact during normal usage.