Trigger QML -> Python log forwarding from the Python side
This should be more robust than assuming that the Python side logging
backend is available once the Python backend initialization finishes.

If the initialization fails for some reason the log output will just
not be redirected and it should be easier to determine what went wrong
because modRana will no longer spam stdout with "logging backend missing"
error messages. :)