Disable logging to stdout in CLI mode
Make sure that the modRana log messages don't pollute the stdout when handling queries in the CLI mode - if the users ask for a
static map URL, only the static map URL should be returned, not also various log messages from modRana startup.

Note: The --debug--startup option can be passed to modRana to keep logging to stdout even when handling CLI queries.