Add show the Route option in the main menu by default
Enable the simple routing mode once it is clicked. In the future it should open a more advanced routing page.
Also remove the routing on/off toggle from the Navigation page in Options - it would still make some sense
for the current simple routing mode, but would not really make sense once the advanced routing page is in place.