Move the Canvas to PinchMap but keep the routing logic in MpaPage
Like this we can make sure that the route is drawn just over the map layer and
under all other elements & we can drop the position indicator hack.
Additionally this should enable easier manipulation of the Canvas during dragging
and adding support for drawing other types of polylines on the map, such as
tracklogs.