Always store the persistent dictionary as version 2 of the marshal module data format
Both Python 2 & 3 can read the version 2 marshal data format. This makes it possible to
run the modRana GTK (Python 2) and Qt 5 GUIs (Python 3) on the same system while sharing
a single options data store.

Note that concurrent usage of the store is not really supported and if two instances of modRana
run at the same time (regardless of GUI type), the one that saves the options last "wins"
and its options state is the one that will persist.