Initial Makefile and spec file
The Makefile should cover most scenarios needed when
building a modRana package for Fedora and Sailfish OS.

The spec file builds the Fedora package by default and
can be set to build a Sailfish package by defining the
with_sailfish macro to equal to 1.

To build a Sailfish Harbour package, both with_sailfish
and with_harbour need to be defined to be equal to 1.

For example:

%define with_sailfish 1
%define with_harbour 1