Add long-click menu with a "route here" button
Add a menu "bubble" that opens when the map is long-clicked.
Just click anywhere on the map and don't move the cursor/finger
and the bubble will show up. The menu currently only contains
a "route here" button that triggers routing from last known good
position to the long-clicked point of the map.