Refactor TurnByTurnPoint
Replace getters and setters by properties and make the code a bit
more PEP8 compliant.

NOTE: This and the other refactorings are meant done now so that
units tests and other planned users of this functionality (as the various
planed CLI interface additions) don't need to change what they call.