Create folder for store if it does not exist yet
Turns out at least in the Sqlite tile store the case of the tile
storage folder not existing was not handled - oops! :P

And while the file based tile store could survive this,
it was still skipping the code needed to assure that
media indexing is disabled for its folder if requested.

Both should be now fixed and the toplevel folder checking & creation
code has been moved to the utils module.