Don't instantiate tile elements until layer list is loaded
This should prevent a nasty race condition that was causing some tiles to be incompletely
initialized only for some layers or no layers at all, resulting in some tiles showing
up as grey squares while other tiles show up fine.

Also as a side effect this might make the map startup a bit faster, as the repeaters don't
need to react on the initial change notifications when the layer overlay list is loaded
and just instantiate the whole list at once.