Update URLs for OpenTopoMap, Hike & Bike and Land/Hill Shading layers
Update URLs for the OpenTopoMap and Hike & Bike layers.

As the OpenPisteMap provided land shading layers basically ceased to exist,
it is in this not really an URL update, but more of a replacement by a
new layer providing similar functionality.

Also thanks to mikehgentry for reporting the issues & helpfully
providing a link to the new URLs! :)

Fixes #134 on GitHub
Fixes #135 on GitHub