Update tiles when panning
First move the tiles model update to a separate thread by running
it with the QML WorkerScript. Like this it will not block the
GUI thread when the tiles model update is running.

It also has a potential benefit of all tile delegate changes happening at once
when the tiles model is synchronized from the thread back to the main QML context.
Previously the changes happened one by one when the tiles model update
code updated the tiles model in the main thread.

Then fix how many tiles we show on the screen - turns out we were
showing one more row & column that needed outside of the visible part of
the map.

Then detect that a new row/column of tiles is about to be shown and
trigger asynchronous tiles model update.

This can result in more tiles model updates & they will happen when the
user is actively dragging the map, so it is important that the update does not
freeze the GUI thread, even for a little while as it would be very
noticeable to the user.

And also set the map background to light gray & draw a lighter gray tile
grid. It turns out that now when tiles are automatically loaded during
panning that it is better if the "background" is not full white but
a less distracting color, such as gray. It also helps if the tile
grid is visible until actual tiles are loaded as this provides feedback
to the user that something is happening & tile images are about to be
loaded.

End result - elimination of yet another potential source of UI freezes
& map that loads new tiles when it is being panned.