Filter logging trace points before drawing
Filter points that are closer than 1.5 meters from the array of points
to draw as the track logging trace on the map.
This should both reduce the amount of weird line drawing artifacts as
well as improve performance by reducing the amount of points that we
need to draw for the trace.

Resolves #172 on GitHub