Add support for map canvas redraw debugging
It is not possible to enable logging of the map cavas redraw time
by toggling an option in Options->Debug. In the future more stuff might
be logged this way.

Also move the canvas wrapping stuff to a separate file called MapCanvas
and add custom paint request functions & paint signals. This both
for wrapping the pain signal for the logging mentioned above but also to
enable retained drawing to the canvas in the future (without clearing
the canvas first).

Resolves #137 on GitHub