Use map coordinates when drawing route and logging traces
Use map coordinates when drawing route and logging traces
as the conversion for map coordinates to screen coordinates *should*
be reasonable cheaper than the conversion from geographic coordinates to
screen coordinates.

It might also make other operations, such as checking if a given point
is too close pixel-wise to draw, doable in the future.

Resolves #174 on GitHub