Add OSM Scout Server local search support
Again enabled from the Options->POI menu.
A point type hint can be specified by prepending a <type:> prefix the search query:

cafe:Steiner

The prefix is likely temporary and a more robust mechanism will be
probably used in the future.

The results (even when using the prefix) still seem to be a little random though. :)