Make top level elements rotate correctly on Sailfish OS
So far toplevel elements, such as the notification popup,
the startup message and debug button have not been
correctly rotating with the rest of the screen content
on device orientation change.

Turns out this was happening due to the Silica ApplicationWindow
not rotating when device orientation changes, unlike the
QtQuick Controls ApplicationWindow which *does* rotate when
the device orientation changes...

So lets just place all the top level content in an Item and
rotate it when needed.

Resolves #191 on GitHub
Resolves #166 on GitHub