Make it possible to set per-layer connection timeout
Add the connection_timeout key for layer in the map
config file, which can be used to set the connection
timeout for tile downloads.

This is mostly needed for offline rendering localhost
tileservers, such as OSM Scout Server, which might
in some cases take some time to render a tile but will
still eventually return it.

Like this we can set a longer connection timeout for
layers provided by localhost renderers while keeping
a short connection timeout for online layers to allow
for quick recovery from connectivity issues.