Hide compass and the clear button during navigation
The compass is in the place of the expected navigation info box
and the clear button is not really needed during navigation.
I't likely also cleaner to first stop the navigation before clearing it.

This also makes it a bit less likely people will clear their navigation route by accident.