Add translation support
Make it possible to translate user visible strings in modRana.

Both QML and Python strings are supported, but as almost nothing is really
marked for translation in Python this is current really relevant only
for QML.

The makefile now has targets:
- pulling string from Python & QML
- pushing new strings to and downloading translations from Transifex
- generate qm files for QML

In general one would just want to run the tx-all target as it does all
that's needed.

Also add localization support to the qt5.sh startup script, even though
translation loading doesn't seem to work for some reason with qmlscene.