Fix ComboBox localization
Turn out the QT_TR_NOOP macro assigns file context to the strings marked for translation. That will not work as the ComboBox implementation is in a different file.

So instead we assume all relevant strings are marker with a QT_TRANSLATE_NOOP macro instead and have a context of "ComboBox".

Fix the Controls and Silica ComboBox implementation and add a note about ComboBox localization to the docs.