Dictionary based about page
Instead of always asynchronously pulling the individual static values at runtime from the Python backend, just grab them at startup and use the resulting dictionary when the page is shown.

This also makes is easier to translate the page as all visible translatable strings can be translated on the QML side. :)