More testing improvements
Make it possible to override the name of the Nosetests executable
(as the Ubuntu package used on Travis has a different binary name from
the one used on Fedora) and make Travis call the makefile for
consistency.

Also set PYTHONPATH during tests to point to the bundle
directory so that tests can mekae use of the bundled dependencies.