Remove the obsolete pyroute and prender libraries
These were some of the initial largely unsuccessful attempts
to have offline routing and map rendering dating back to the
Rana and Neo FreeRunner times.

Now we finally have very robust, fast and full featured offline
routing and rendering (and much more) in the form of OSM Scout Server
and pyrender & pyrender never really worked after all.

So get rid of those and yet again - we can always dig them out of the
history in the unlikely case the code is ever needed for something
again.