Add the voice module, originally from whogo-maps
Thanks a lot to rinigius and otsaloma for creating such
a clean and portable multi-backend TTS handling module! :)

Thanks to this it will be much easier to support not just
Espeak but also all the other TTS backends rinigius ported
to Sailfish OS. Also thanks to the WAV file generation method
TTS sound playback should be in the future totally seamless
for users and just installing one of the TTS backends will
be enough.