Port tests for the voice module from WhoGo maps
Port the tests to be a UnitTest subclass and skip the individual engine
tests for now. We will likely want to split that method in the future to
test the various backends individually and based on their availability.
That would also mean configuring Travis accordingly so it won't blow up
due to missing TTS engine binaries.

Also we might want to do something about the sleep() calls it tests, as they can potentially make the tests quite long - either the sample creation calls could be synchronized somehow or the tests run in parallel (IIRC that's supported by the Python unittest framework or Nose).