Run tests on latest Fedora via docker in Travis
As Travis is still stuck on Ubuntu 14.04 (!?) as the latest available
stable build environment let's use a temporary Docker container to
pull in the latest stable Fedora image and run the Python unit tests on
that. That should make the test result also more relevant, rather than
testing on ancient distro that will soon go out of support.

BTW, this is about 100x times less efficient than just running the
tests on a sane modern distro but it can't be helped!

We only run thy Python unit tests as qmltestrunner apparently
segfaults (?!) itself or even the container when we try to run it
inside a container. This is funny as missing PyOtherSide needed by the
QML unit tests was the original reason for migrating the tests
to a run on a Fedora container. :D

Also thanks a lot to the Lorax project on which this this Docker based
testing methodology is based.