Rework how we pass speed statistics to QML
Instead of passing finished strings with units,
pass rounded integer values corresponding to
current/average/max speed.

This way the units can be added on the QML side
and properly localized. Also it open the possibility
of using the value for other purposes than just
displaying them, such as for example driving
and animated speedometer element.