Remove the getMap() function
The getMap() function was used to directly access the underlying PinchMap
elements of the map page. As all direct calls have been replaced by
calls that get through the BaseMapPage API, direct access is no longer
needed and the getMap() function can be safely removed.

This is actually another step towards supporting the MapBox GL Native map
widget as it of course does not contain any PinchMap and instead its
functionality will be available in a common way via the BaseMapPage API.