Add the TwoOptionPage element
The main purpose of this page is to serve basically
as a confirmation dialog in case where confirmation
of a decision is required.

Of course, there is the remorse timer on Sailfish OS,
but I never really liked the concept:
- easy for the remorse timer to get garbage collected
  when you leave a page
- you have to wait for it to finish even if you are sure you
  want to complete the action (if basically adds a fixed
  delay to UI workflow)

Not to mention the element is Silica-only and would have to
be reimplemented with Controls in UC to be usable for modRana.

So let's go with a normal yes/no page, like everyone else. ;-)